Added superclass checking for the signalWithReceptionInClass pattern
To reduce the number of dependencies between pattern files, some pattern
were moved between eiq files; however, as they are still in the same
Java package, no user of these patterns should be affected.

This is supposed to fix the issue #66 and also an issue with the UML
transformation as well.